# config/sofair-openapi.yml
components:
...
  responses:
    '200':
      description: successful operation
...
    default:
      content:
        application/json:
          schema:
            $ref: https://schemas.opengis.net/ogcapi/processes/part1/1.0/openapi/schemas/exception.yaml
      description: Unexpected error
...
info:
  contact:
    email: geosas_adm@framalistes.org
    name: "G\xE9oSAS"
    url: https://geosas.fr
  description: pygeoapi provid
  license:
    name: CC-BY 4.0 license
    url: https://creativecommons.org/licenses/by/4.0/
  termsOfService: https://creativecommons.org/licenses/by/4.0/
  title: SOFAIR API
  version: 0.16.dev0
  x-keywords:
  - SensorThings
  - OGC
  - INSPIRE
openapi: 3.0.2
paths:
...
  /processes/create-config-xlsx/execution:
    post:
      description: 'This function uploads an XLSX file to create the various objects
        in the SensorThings service: Things, Sensors, Observed Properties, Locations,
        Features Of Interest and Datastreams.'
      operationId: executeCreate-config-xlsxJob
      requestBody:
        content:
          application/json:
            example:
              inputs:
                config_xlsx: aghrys_sensorthings.xlsx
            schema:
              $ref: https://schemas.opengis.net/ogcapi/processes/part1/1.0/openapi/schemas/execute.yaml
        description: Mandatory execute request JSON
        required: true
      responses:
        '200':
          $ref: '#/components/responses/200'
        '201':
          $ref: https://schemas.opengis.net/ogcapi/processes/part1/1.0/openapi/responses/ExecuteAsync.yaml
        '404':
          $ref: https://schemas.opengis.net/ogcapi/processes/part1/1.0/openapi/responses/NotFound.yaml
        '500':
          $ref: https://schemas.opengis.net/ogcapi/processes/part1/1.0/openapi/responses/ServerError.yaml
        default:
          $ref: '#/components/responses/default'
      summary: Process Setting up the sensorthings service execution
      tags:
      - create-config-xlsx
...
  /processes/post-observations/execution:
    post:
      description: This function uploads observations from CSV or XLSX file to the
        SensorThings service.
      operationId: executePost-observationsJob
      requestBody:
        content:
          application/json:
            example:
              inputs:
                colonne_date: date
                data_file: meteo_sensorthings.csv
                format_date: '%Y%m%d'
                mode: large
            schema:
              $ref: https://schemas.opengis.net/ogcapi/processes/part1/1.0/openapi/schemas/execute.yaml
        description: Mandatory execute request JSON
        required: true
      responses:
        '200':
          $ref: '#/components/responses/200'
        '201':
          $ref: https://schemas.opengis.net/ogcapi/processes/part1/1.0/openapi/responses/ExecuteAsync.yaml
        '404':
          $ref: https://schemas.opengis.net/ogcapi/processes/part1/1.0/openapi/responses/NotFound.yaml
        '500':
          $ref: https://schemas.opengis.net/ogcapi/processes/part1/1.0/openapi/responses/ServerError.yaml
        default:
          $ref: '#/components/responses/default'
      summary: Process Upload observations execution
      tags:
      - post-observations
...
servers:
- description: pygeoapi provid
  url: https://geosas.fr/sofair-dev/api
...